requirements.txt
flask==2.2.3
flask-cors==3.0.10
pymongo==4.3.3
python-dotenv==1.0.0
requests==2.28.2
google-generativeai==0.3.1
numpy
pypdf==3.15.1
Werkzeug==2.2.3
